Fix assignment and remove delete and balance
Originally, I used Node objects instead of values to build the binary
search trees. I fixed that. Also, because I have a decent grasp of what
is going on, and I just want to get this graded, I did not spend the
time to fix the delete and tree balancing. So I removed them because
they are not required.